// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated by Convex
 * This file is generated and should not be manually edited.
 */
import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";
import type * as activities from "../activities.js";
import type * as scheduledTasks from "../scheduledTasks.js";
import type * as search from "../search.js";

declare const fullApi: ApiFromModules<{
  activities: typeof activities;
  scheduledTasks: typeof scheduledTasks;
  search: typeof search;
}>;

export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;